refactor(steps): some updates
