+++ src/simple-todos/step02/imports/ui/App.vue
@@ -1,7 +1,7 @@
 <script setup>
 import Task from './components/Task.vue';
 import { subscribe, autorun } from 'vue-meteor-tracker';
-import { TasksCollection } from '../api/TasksCollection';
+import { TasksCollection } from '../db/TasksCollection';
 
 subscribe('tasks');
 const tasks = autorun(() => TasksCollection.find({}).fetch()).result;
