+++ src/simple-todos/step02/server/main.js
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import { TasksCollection } from '../imports/api/TasksCollection';
+import { TasksCollection } from '../imports/db/TasksCollection';
 import '../imports/api/tasksPublications';
 
 const insertTask = (taskText) => TasksCollection.insert({ text: taskText });
