+++ src/simple-todos/step03/imports/api/tasksPublications.js
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import { TasksCollection } from './TasksCollection';
+import { TasksCollection } from '../db/TasksCollection';
 
 Meteor.publish('tasks', function publishTasks() {
   return TasksCollection.find({ userId: this.userId });
