+++ src/simple-todos/step05/imports/api/tasksMethods.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { check } from 'meteor/check';
-import { TasksCollection } from '../api/TasksCollection';
+import { TasksCollection } from '../db/TasksCollection';
 
 Meteor.methods({
   'tasks.insert'(text) {
