+++ src/simple-todos/step06/imports/ui/App.vue
@@ -3,7 +3,7 @@ import { ref } from 'vue';
 import Task from './components/Task.vue';
 import TaskForm from './components/TaskForm.vue';
 import { subscribe, autorun } from 'vue-meteor-tracker';
-import { TasksCollection } from '../api/TasksCollection';
+import { TasksCollection } from '../db/TasksCollection';
 
 const hideCompleted = ref(false);
