+++ src/simple-todos/step07/imports/ui/App.vue
@@ -6,7 +6,7 @@ import Task from './components/Task.vue';
 import TaskForm from './components/TaskForm.vue';
 import LoginForm from './components/LoginForm.vue';
 import { subscribe, autorun } from 'vue-meteor-tracker';
-import { TasksCollection } from '../api/TasksCollection';
+import { TasksCollection } from '../db/TasksCollection';
 
 const hideCompleted = ref(false);
 const isLogged = ref(false);
